move styles into seperate file
